fix: move cip509 validation
